src/requirements.txt
azure-ai-inference[opentelemetry]
azure.cosmos
azure-identity
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
azure-storage-blob
Flask
langchain
langchain-community
langchain-openai
langgraph
openinference.instrumentation.langchain
python-dotenv
azure-monitor-opentelemetry
azure-identity
opentelemetry-sdk 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-flask 
azure-monitor-opentelemetry-exporter
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
opentelemetry-sdk
python-dotenv
